Also pass the document object to the handler.
In the case where no nodes match an XPath, the NodeList is empty. This
would allow a handler to throw an exception if it required a match, but
it gives no access to the document if the handler wants to perform some
other action, such as add a new node. Passing the document (which is
available anyway when the NodeList is not empty) solves this issue for
the empty list case.